♻️ Refactor music blueprint to be more modular & maintainable
