+++ server/blueprints/music/__init__.py
@@ -1,0 +1,8 @@
+from colorama import Fore
+
+from utils.logging import Logger
+from utils.subsonic import Subsonic
+
+subsonic = Subsonic()
+
+logger = Logger("Music", Fore.LIGHTCYAN_EX)
