+++ server/utils/logging.py
@@ -1,15 +1,18 @@
-from typing import Callable
-
-from colorama import Fore, Style
+from colorama import Style
 
 from .messages import info, warn, error
 
-print_info: Callable[[str], None] = lambda message: info(
-    f"{Fore.BLUE}Rhea{Style.RESET_ALL} ", message
-)
-print_warn: Callable[[str], None] = lambda message: warn(
-    f"{Fore.BLUE}Rhea{Style.RESET_ALL} ", message
-)
-print_error: Callable[[str], None] = lambda message: error(
-    f"{Fore.BLUE}Rhea{Style.RESET_ALL} ", message
-)
+
+class Logger:
+    def __init__(self, name: str, color: str):
+        self.name = name
+        self.color = color
+
+    def info(self, message: str) -> None:
+        info(f"{self.color}{self.name}{Style.RESET_ALL} ", message)
+
+    def warn(self, message: str) -> None:
+        warn(f"{self.color}{self.name}{Style.RESET_ALL} ", message)
+
+    def error(self, message: str) -> None:
+        error(f"{self.color}{self.name}{Style.RESET_ALL} ", message)
